check

+++ train.py
@@ -16,23 +16,24 @@ from model.model import *
 from attack.attack import *
 from sklearn.metrics import roc_auc_score
 
-print(f"Centralized: {args.data} --------------------------------------")
-strategy, AUTO = getStrategy()
-with strategy.scope():
-    model, preprocess = model_factory()
-    model.compile(optimizer = tf.keras.optimizers.SGD(learning_rate=1e-2),
-                    loss = {'output': tf.keras.losses.SparseCategoricalCrossentropy()},
-                    metrics = {"output": [tf.keras.metrics.SparseCategoricalAccuracy()]})
+if not os.path.isfile(f'{data.__FOLDER__}/cen_{args.data}.pickle'):
+    print(f"Centralized: {args.data} --------------------------------------")
+    strategy, AUTO = getStrategy()
+    with strategy.scope():
+        model, preprocess = model_factory()
+        model.compile(optimizer = tf.keras.optimizers.SGD(learning_rate=1e-2),
+                        loss = {'output': tf.keras.losses.SparseCategoricalCrossentropy()},
+                        metrics = {"output": [tf.keras.metrics.SparseCategoricalAccuracy()]})
 
-trainData, inOutLabels = data.loadCenData(args.data, preprocess)
-model.fit(trainData, verbose = args.verbose, epochs = 100)
+    trainData, inOutLabels = data.loadCenData(args.data, preprocess)
+    model.fit(trainData, verbose = args.verbose, epochs = 100)
 
-miaData, validData = data.load(preprocess)
-validPred = model.predict(validData, verbose = args.verbose)
-print("Validation:", np.mean(validData.labels.flatten() == np.argmax(validPred, axis = 1)))
+    miaData, validData = data.load(preprocess)
+    validPred = model.predict(validData, verbose = args.verbose)
+    print("Validation:", np.mean(validData.labels.flatten() == np.argmax(validPred, axis = 1)))
 
-MIAPred = model.predict(miaData, verbose = args.verbose)
-print("MIA:", np.mean(miaData.labels.flatten() == np.argmax(MIAPred, axis = 1)))
+    MIAPred = model.predict(miaData, verbose = args.verbose)
+    print("MIA:", np.mean(miaData.labels.flatten() == np.argmax(MIAPred, axis = 1)))
 
-with open(f'{data.__FOLDER__}/cen_{args.data}.pickle', 'wb') as handle:
-    pickle.dump([MIAPred, inOutLabels], handle, protocol=pickle.HIGHEST_PROTOCOL)
+    with open(f'{data.__FOLDER__}/cen_{args.data}.pickle', 'wb') as handle:
+        pickle.dump([MIAPred, inOutLabels], handle, protocol=pickle.HIGHEST_PROTOCOL)
